feat: delete search.ts → browser-direct global search + bridge the agent tool (WS2)
Global search re-homed browser-direct: convex/globalSearch.ts search guard relaxed
requireService→requireOrgRead (service tokens still allowed for PDF/API); new
src/hooks/use-global-search.ts (imperative useConvex().query, 2-char min, identical
{results} shape); command-search.tsx rewired (with globalSearch in the effect deps —
codex caught a stale-closure bug where org resolving post-mount would never retry).
SearchResult/SearchResultType moved to src/lib/search-types.ts. src/server/search.ts
DELETED. The `global_search` agent tool is PRESERVED (not degraded) by exposing
api.globalSearch.search via the CONVEX_READS bridge under the same op name + scope
(asset:read) — the cheap re-point path for a core tool.
